requirements.txt
requests
python-allrecipes
